feat: Refactor WebGLNode and WebRetrievalNode components to use composable functions for improved structure and reusability
