--- frontend/src/components/WebRetrievalNode.vue.d.ts
@@ -1,23 +1,0 @@
-// e.g. WebRetrievalNode.vue.d.ts
-export interface WebRetrievalNodeData {
-  type: string
-  labelStyle?: Record<string, any>
-  style?: Record<string, any>
-  inputs?: {
-    url?: string
-  },
-  outputs?: {
-    // reflect the new property
-    result?: {
-      output?: string
-    }
-  },
-  hasInputs?: boolean
-  hasOutputs?: boolean
-  inputHandleColor?: string
-  inputHandleShape?: string
-  handleColor?: string
-  outputHandleShape?: string
-  run?: () => Promise<any>
-  error?: string
-}
